@@ -1,13 +1,13 @@
 import { useQuery } from '@apollo/client';
 import React from 'react';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router-dom';
 import { Button, Flex, Image, Text } from 'rebass';
-import { Spinner } from '..';
+import { HomeButton, Spinner } from '..';
 import { AUTH_QUERY } from '../../gql';
 import { Container } from '../Container';
 import { TutorialHeader } from '../TutorialHeader';
 import logo from './../../assets/logoskoly.png';
-import { SlideProps } from './types';
+import { FormattedMessage } from 'react-intl';
 
 const imageStyle = {
   minHeight: '210px',
@@ -18,47 +18,47 @@ const imageStyle = {
   py: 3,
 };
 
-const Slide15 = (props: SlideProps) => {
+const Slide15 = () => {
   const { loading, data } = useQuery(AUTH_QUERY, {
     fetchPolicy: 'cache-first',
   });
 
   const isLoggedIn = !!data?.viewer?.user;
-  const isAdmin = data?.viewer?.user?.role === 'admin'; 
+  const isAdmin = data?.viewer?.user?.role === 'admin';
 
-  const history = useHistory();
-  const handleClickHome = () => {
-    history.push('/');
-  };
+  const navigate = useNavigate();
   const handleClickPlay = () => {
-    history.push('/play');
+    navigate('/play');
   };
-  const handleClickSignIn = () => {
-    history.push('/signin');
+  const handleClickSignUp = () => {
+    navigate('/signup');
   };
 
   return (
     <Container>
       <TutorialHeader>
         <Text fontSize={[4, 4, 5]} color="secondary" textAlign="center" p={4}>
-          To je vše, nyní jsi připraven*a na hru naostro!
+          <FormattedMessage
+            id="app.tutorial.slide.ready"
+            defaultMessage="That's it! You're now ready to play for real!"
+            description="Tut15 ready"
+          />
         </Text>
       </TutorialHeader>
       <Image src={logo} sx={imageStyle} />
       <Text fontSize={[2, 3, 4]} color="secondary" textAlign="center" p={3}>
-        Tip for Science bylo vytvořeno evolučními biology
-        <br />
-        z Univerzity Karlovy.
+        <FormattedMessage
+          id="app.tutorial.slide.byeb"
+          defaultMessage="Tip for Science was created by evolutionary biologists {lineBreak}
+            from Charles University."
+          description="Tut15 by EB"
+          values={{
+            lineBreak: <br />,
+          }}
+        />
       </Text>
       <Flex marginTop={'auto'}>
-        <Button
-          onClick={handleClickHome}
-          sx={{ flex: 1 }}
-          mr={1}
-          backgroundColor={'#414141'}
-        >
-          Domů
-        </Button>
+        <HomeButton />
         {loading && (
           <Flex justifyContent="center">
             <Spinner />
@@ -66,12 +66,20 @@ const Slide15 = (props: SlideProps) => {
         )}
         {!loading && isLoggedIn && !isAdmin && (
           <Button onClick={handleClickPlay} sx={{ flex: 3 }}>
-            Začít hrát
+            <FormattedMessage
+              id="app.tutorial.slide.play"
+              defaultMessage="Play"
+              description="Tut15 play"
+            />
           </Button>
         )}
         {!loading && !isLoggedIn && (
-          <Button onClick={handleClickSignIn} sx={{ flex: 3 }}>
-            Přihlásit se
+          <Button onClick={handleClickSignUp} sx={{ flex: 3 }}>
+            <FormattedMessage
+              id="app.tutorial.slide.signin"
+              defaultMessage="Sign in"
+              description="Tut15 signin"
+            />
           </Button>
         )}
       </Flex>
