@@ -1,14 +1,17 @@
 import React, { useState } from 'react';
-import { Flex, Button, Box, Text } from 'rebass';
+import { Flex, Box, Text, Heading } from 'rebass';
 import { Label, Input } from '@rebass/forms';
 import { useForm } from 'react-hook-form';
-
-import { Container } from '../../components';
-import { Link, useHistory } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
+import { FormattedMessage, useIntl } from 'react-intl';
+import { MenuButton, Container } from '../../components';
 import { useMutation } from '@apollo/client';
 import { AuthQueryName, SIGN_IN_MUTATION } from '../../gql';
 import Helmet from 'react-helmet';
 
+const buttonStyles = {
+  my: 3,
+};
 const inputStyles = {
   '::placeholder': {
     color: 'white',
@@ -22,10 +25,16 @@ const labelStyles = {
 };
 
 const SignIn = () => {
+  const intl = useIntl();
+  const helmet = intl.formatMessage({
+    id: 'app.sigin',
+    defaultMessage: `Sign in`,
+    description: 'Sign in',
+  });
   const { register, handleSubmit } = useForm();
 
   const [errors, setErrors] = useState([]);
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const [signIn] = useMutation(SIGN_IN_MUTATION, {
     refetchQueries: [AuthQueryName],
@@ -44,7 +53,7 @@ const SignIn = () => {
     });
 
     if (data && !data.signIn.errors) {
-      history.push('/');
+      navigate('/');
     } else {
       setErrors(data!.signIn!.errors);
     }
@@ -52,61 +61,110 @@ const SignIn = () => {
 
   return (
     <Container>
-      <Helmet title="Přihlášení"></Helmet>
+      <Helmet title={helmet}></Helmet>
       <Flex
-        as="form"
-        onSubmit={handleSubmit(onSubmit)}
         flexDirection="column"
         justifyContent="center"
         height="100%"
         width="100%"
         p="3"
       >
-        <Label htmlFor="email" sx={labelStyles}>
-          E-mail
-        </Label>
-        <Input
-          id="email"
-          type="text"
-          placeholder="e-mail"
-          mb={2}
-          sx={inputStyles}
-          {...register('email')}
-        />
-        <Label htmlFor="password" sx={labelStyles}>
-          Heslo
-        </Label>
-        <Input
-          id="password"
-          type="password"
-          mb={2}
-          sx={inputStyles}
-          {...register('password')}
-        />
-        <Button type="submit" my={3}>
-          Přihlásit
-        </Button>
-        {errors.length > 0 && (
-          <Box>
-            {errors.map((item: { error: string }) => (
-              <Text color="red">{item.error}</Text>
-            ))}
-          </Box>
-        )}
+        <Flex
+          sx={{ flexGrow: 2 }}
+          justifyContent="center"
+          flexDirection="column"
+        >
+          <Heading
+            textAlign="left"
+            color="#D6D6D6"
+            fontWeight={900}
+            lineHeight="0.85em"
+            mx="auto"
+            fontSize={80}
+          >
+            TIP FOR <br />
+            SCIENCE
+          </Heading>
+        </Flex>
         <Flex justifyContent="center">
           <Text color="white" fontSize="1">
-            Nemáš účet?{' '}
+            <FormattedMessage
+              id="app.signin.menu.noacc"
+              defaultMessage="No account yet?"
+              description="SignIn NoAcc text"
+            />{' '}
             <Link to="/signup" style={{ color: '#FF0070' }}>
-              Vytvořit účet
+              <FormattedMessage
+                id="app.signin.menu.create"
+                defaultMessage="Create an account"
+                description="SignIn CreateAcc link"
+              />
             </Link>
           </Text>
         </Flex>
-        <Flex justifyContent="center" my="2">
-          <Text color="white" fontSize="1">
-            <Link to="/" style={{ color: '#D76B90' }}>
-              Domů
-            </Link>
-          </Text>
+        <Flex
+          as="form"
+          onSubmit={handleSubmit(onSubmit)}
+          flexDirection="column"
+          justifyContent="center"
+          width="100%"
+          p="1"
+        >
+          <Label htmlFor="email" sx={labelStyles}>
+            <FormattedMessage
+              id="app.signin.menu.email"
+              defaultMessage="E-mail"
+              description="SignIn E-mail label"
+            />
+          </Label>
+          <Input
+            id="email"
+            type="text"
+            placeholder="e-mail"
+            mb={2}
+            sx={inputStyles}
+            {...register('email')}
+          />
+          <Label htmlFor="password" sx={labelStyles}>
+            <FormattedMessage
+              id="app.signin.menu.password"
+              defaultMessage="Password"
+              description="SignIn Password label"
+            />
+          </Label>
+          <Input
+            id="password"
+            type="password"
+            mb={2}
+            sx={inputStyles}
+            {...register('password')}
+          />
+          <MenuButton type="submit" sx={buttonStyles} primary>
+            <FormattedMessage
+              id="app.signin.menu.login"
+              defaultMessage="Login"
+              description="SignIn Login button"
+            />
+          </MenuButton>
+          {errors.length > 0 && (
+            <Box>
+              {errors.map((item: { error: string }) => (
+                <Text color="red">{item.error}</Text>
+              ))}
+            </Box>
+          )}
+
+          <Flex justifyContent="center" my="2">
+            <Text color="white" fontSize="1">
+              <Link to="/" style={{ color: '#D76B90' }}>
+                <FormattedMessage
+                  id="app.signin.menu.home"
+                  defaultMessage="Home"
+                  description="SignIn Home link"
+                />
+              </Link>
+            </Text>
+          </Flex>
         </Flex>
       </Flex>
     </Container>
