@@ -1,6 +1,7 @@
 import React from 'react';
 import { Text, Flex } from 'rebass';
 import NumberFormat from 'react-number-format';
+import { FormattedMessage } from 'react-intl';
 
 interface PreviousTipsProps {
   previousTips?: number[];
@@ -8,7 +9,7 @@ interface PreviousTipsProps {
 }
 
 const previousTipStyle = {
-  background: '#FF0070',
+  background: '#e91e63',
   mr: 1,
   px: 2,
   py: 1,
@@ -18,7 +19,11 @@ const PreviousTips = ({ previousTips, unit }: PreviousTipsProps) =>
   previousTips != null && previousTips.length !== 0 ? (
     <>
       <Text textAlign="center" color="#FF0070" my={1}>
-        Předchozí tipy:{' '}
+        <FormattedMessage
+          id="app.previoustis"
+          defaultMessage="Previous tips: "
+          description="Previous tips"
+        />
       </Text>
       <Flex justifyContent="center">
         {previousTips.map((previousTip, index) => (
