@@ -1,7 +1,9 @@
 import React from 'react';
-import { Button, Flex, Text } from 'rebass';
+import { Flex, Text } from 'rebass';
 import { Container } from '../Container';
 import { TranslucentBox } from '../TranslucentBox';
+import Button from '@mui/material/Button';
+import { FormattedMessage } from 'react-intl';
 
 const TooCloseDialog = ({
   onGuessed,
@@ -29,15 +31,33 @@ const TooCloseDialog = ({
           height="100%"
         >
           <Text fontWeight={500} textAlign="center" my={3} color="white">
-            Wow! To bylo FAKT blízko! Věděl*a jsi správnou odpověď, nebo se ti
-            opravdu podařilo takhle dobře tipnout?
+            <FormattedMessage
+              id="app.game.tooclose"
+              defaultMessage="Wow! That was REALLY close! Did you actually know the correct answer, or did you just make an amazing guess?"
+              description="Too close"
+            />
             <br />
-            Neboj, tvé skóre to neovlivní, ale pomůže nám to při vyhodnocování
-            dat :)
+            <FormattedMessage
+              id="app.game.closedontworry"
+              defaultMessage="Don't worry, this won't affect your score, but it helps us with data analysis. :)"
+              description="Don't worry"
+            />
           </Text>
           <Flex justifyContent="space-between" width="100%">
-            <Button onClick={handleClickGuessed}>Tipnul*a jsem</Button>
-            <Button onClick={handleClickKnewIt}> Věděl*a jsem</Button>
+            <Button onClick={handleClickGuessed}>
+              <FormattedMessage
+                id="app.game.guessed"
+                defaultMessage="I guessed"
+                description="Guessed"
+              />
+            </Button>
+            <Button onClick={handleClickKnewIt}>
+              <FormattedMessage
+                id="app.game.knew"
+                defaultMessage="I knew"
+                description="Knew"
+              />
+            </Button>
           </Flex>
         </Flex>
       </Container>
