@@ -1,17 +1,12 @@
 import React from 'react';
-import { useHistory } from 'react-router-dom';
-import { Button, Text, Box, Flex } from 'rebass';
+import { Text, Box, Flex } from 'rebass';
 import { Container } from '../Container';
 import { SlideProps } from './types';
+import { HomeButton } from './HomeButton';
+import { NextButton } from './NextButton';
+import { FormattedMessage } from 'react-intl';
 
 const Slide11 = ({ handleNextStep }: SlideProps) => {
-  const history = useHistory();
-  const handleClickHome = () => {
-    history.push('/');
-  };
-  const handleClickNext = () => {
-    handleNextStep();
-  };
   return (
     <Container>
       <Box p="4" my="auto">
@@ -22,28 +17,30 @@ const Slide11 = ({ handleNextStep }: SlideProps) => {
           mb="4"
           mt="4"
         >
-          Vypilovali jsme sociální přenos téměř k dokonalosti.
+          <FormattedMessage
+            id="app.tutorial.slide.perfected"
+            defaultMessage="We have almost perfected social transmission."
+            description="Tut11 perfected"
+          />
         </Text>
         <Text fontSize={[3, 4, 5]} color="secondary" textAlign="center" mb="4">
-          Učíme se jeden od druhého - tvoříme, stavíme, malujeme, píšeme básně -
-          inspirováni malířkami, vynálezci, vědkyněmi a básníky minulosti.
+          <FormattedMessage
+            id="app.tutorial.slide.poets"
+            defaultMessage="We learn from each other - we create, build, paint, write poems - inspired by painters, inventors, scientists and poets of the past."
+            description="Tut11 poets"
+          />
         </Text>
         <Text fontSize={[3, 4, 5]} color="secondary" textAlign="center" mb="4">
-          Využíváme výhod tzv. kulturní dědičnosti.
+          <FormattedMessage
+            id="app.tutorial.slide.inheritance"
+            defaultMessage="We take advantage of the cultural inheritance."
+            description="Tut11 inheritance"
+          />
         </Text>
       </Box>
       <Flex mt="auto" justifyContent="space-between" width="100%">
-        <Button
-          onClick={handleClickHome}
-          backgroundColor={'#414141'}
-          sx={{ flex: 1 }}
-          mr="1"
-        >
-          Domů
-        </Button>
-        <Button onClick={handleClickNext} sx={{ flex: 5 }}>
-          Další
-        </Button>
+        <HomeButton />
+        <NextButton handleNextStep={handleNextStep} />
       </Flex>
     </Container>
   );
