@@ -1,17 +1,12 @@
 import React from 'react';
-import { useHistory } from 'react-router-dom';
-import { Button, Text, Box, Flex } from 'rebass';
+import { Text, Box, Flex } from 'rebass';
 import { Container } from '../Container';
 import { SlideProps } from './types';
+import { HomeButton } from './HomeButton';
+import { NextButton } from './NextButton';
+import { FormattedMessage } from 'react-intl';
 
 const Slide12 = ({ handleNextStep }: SlideProps) => {
-  const history = useHistory();
-  const handleClickHome = () => {
-    history.push('/');
-  };
-  const handleClickNext = () => {
-    handleNextStep();
-  };
   return (
     <Container>
       <Box p="4" my="auto">
@@ -22,23 +17,18 @@ const Slide12 = ({ handleNextStep }: SlideProps) => {
           mb="4"
           mt="4"
         >
-          Tato hra je jeden velký experiment v rámci kulturní evoluce. Zajímá
-          nás, jak tento druh extrémně spolupracujících opic nalezne optimální
-          strategii, když bude mít k dispozici pouze skupinovou inteligenci.
+          <FormattedMessage
+            id="app.tutorial.slide.apes"
+            defaultMessage="This game is one big experiment in cultural evolution. Interested
+            us how this species of extremely cooperative apes finds the optimum
+            strategy when only group intelligence is available."
+            description="Tut12 apes"
+          />
         </Text>
       </Box>
       <Flex mt="auto" justifyContent="space-between" width="100%">
-        <Button
-          onClick={handleClickHome}
-          backgroundColor={'#414141'}
-          sx={{ flex: 1 }}
-          mr="1"
-        >
-          Domů
-        </Button>
-        <Button onClick={handleClickNext} sx={{ flex: 5 }}>
-          Další
-        </Button>
+        <HomeButton />
+        <NextButton handleNextStep={handleNextStep} />
       </Flex>
     </Container>
   );
