@@ -7,6 +7,7 @@ import { Container } from '../Container';
 import { TutorialHeader } from '../TutorialHeader';
 import { SlideProps } from './types';
 import { SubmitButton } from '../SubmitButton';
+import { FormattedMessage, useIntl } from 'react-intl';
 
 const imageStyle = {
   maxWidth: '100%',
@@ -38,12 +39,27 @@ const labelStyle = {
   color: 'white',
 };
 
-const question = 'Jak rychle se otáčí Jupiter? A pozor, běží ti čas!';
-const unit = 'km/h';
 const previousTips = [80000, 1000, 500, 20000];
 const timeLimit = 20;
 
 const Slide13 = ({ onSubmit }: SlideProps) => {
+  const intl = useIntl();
+  const question = intl.formatMessage({
+    id: 'app.tutorial.slide.jupiter',
+    defaultMessage: `How fast does Jupiter rotate? And hurry up, time is ticking!`,
+    description: 'Tut13 jupiter',
+  });
+  const unit = intl.formatMessage({
+    id: 'app.unitkmh',
+    defaultMessage: `km/h`,
+    description: 'unit km/h',
+  });
+  const placeholder = intl.formatMessage({
+    id: 'app.yourtip',
+    defaultMessage: 'your tip',
+    description: 'Yourtip placeholder',
+  });
+
   const inputRef = useRef<HTMLInputElement>(null);
   const handleClickSubmit = () => {
     onSubmit(Number(inputRef.current!.value));
@@ -103,13 +119,17 @@ const Slide13 = ({ onSubmit }: SlideProps) => {
       <PreviousTips previousTips={previousTips} unit={unit} />
       <Flex justifyContent="center" alignItems="center" p={2}>
         <Label htmlFor="tip" sx={labelStyle}>
-          tip:
+          <FormattedMessage
+            id="app.tip"
+            defaultMessage="tip:"
+            description="Tip"
+          />
         </Label>
         <Input
           id="tip"
           name="tip"
           type="number"
-          placeholder="tvůj tip"
+          placeholder={placeholder}
           sx={inputStyles}
           onKeyDown={handleSubmit}
           ref={inputRef}
@@ -133,7 +153,11 @@ const Slide13 = ({ onSubmit }: SlideProps) => {
           }}
         >
           <Text fontWeight={500} textAlign="center" my={1} color="white">
-            Vypršel ti čas! Tentokrát ale ještě můžeš zadat odpověď zde:
+            <FormattedMessage
+              id="app.tutorial.slide.timesup"
+              defaultMessage="Time’s up! But this time, you can still submit your answer here:"
+              description="Tut13 times up"
+            />
           </Text>
           <Input
             id="tip2"
