@@ -1,31 +1,25 @@
 import React from 'react';
-import { Button, Box, Text, Flex } from 'rebass';
-import { useHistory } from 'react-router';
+import { Box, Text, Flex } from 'rebass';
 import { Container } from '../Container';
 import { TutorialHeader } from '../TutorialHeader';
 import elephant from './../../assets/elephantTut.jpg';
 import { SlideProps } from './types';
-import {
-  getScore,
-  getScoreSentence,
-  topScoreSentence,
-  highScoreSentence,
-  lowScoreSentence,
-  zeroScoreSentence,
-} from '../../helpers';
+import { getScore } from '../../helpers';
 import { getTutorialImageStyle } from '../commonStyleSheets';
+import { HomeButton } from './HomeButton';
+import { NextButton } from './NextButton';
+import { FormattedMessage, useIntl } from 'react-intl';
+import ScoreMessage from '../ScoreMessage';
 
 const imageStyle = getTutorialImageStyle(elephant);
 
 const Slide4 = ({ handleNextStep, currentTip }: SlideProps) => {
-  const history = useHistory();
-  const handleClickHome = () => {
-    history.push('/');
-  };
-  const handleClickNext = () => {
-    handleNextStep();
-  };
-
+  const intl = useIntl();
+  const unit = intl.formatMessage({
+    id: 'app.unitkg',
+    defaultMessage: 'kg',
+    description: 'unit kg',
+  });
   const questionScore = getScore(
     typeof currentTip === 'undefined' ? 0 : currentTip,
     10886
@@ -34,21 +28,18 @@ const Slide4 = ({ handleNextStep, currentTip }: SlideProps) => {
     <Container>
       <TutorialHeader>
         <Text fontSize={[3, 4, 5]} color="accent" textAlign="center" p={2}>
-          {currentTip} kg?{' '}
+          {currentTip} {unit}?{' '}
           <Text color="secondary" as="span">
-            {questionScore === 0 && getScoreSentence(zeroScoreSentence)}
+            {questionScore === 0 && <ScoreMessage scoreType="score.zero" />}
             {questionScore !== null &&
               questionScore > 0 &&
-              questionScore < 40 &&
-              getScoreSentence(lowScoreSentence)}
+              questionScore < 40 && <ScoreMessage scoreType="score.low" />}
             {questionScore !== null &&
               questionScore >= 40 &&
-              questionScore < 80 &&
-              getScoreSentence(highScoreSentence)}
+              questionScore < 80 && <ScoreMessage scoreType="score.high" />}
             {questionScore !== null &&
               questionScore >= 80 &&
-              questionScore < 95 &&
-              getScoreSentence(topScoreSentence)}
+              questionScore < 95 && <ScoreMessage scoreType="score.top" />}
           </Text>
         </Text>
       </TutorialHeader>
@@ -60,25 +51,23 @@ const Slide4 = ({ handleNextStep, currentTip }: SlideProps) => {
         py={4}
         px={3}
       >
-        Největší samec slona, jehož váhu se podařilo zaznamenat, vážil{' '}
+        <FormattedMessage
+          id="app.tutorial.slide.elemale"
+          defaultMessage="The largest male elephant on record weighed "
+          description="Tut4 elemale"
+        />
         <Text color="white" as="span">
           10 886
         </Text>
-         kg a v kohoutku měřil 3,96 metru. Byl tedy téměř o metr vyšší než
-        průměrný slon africký.
+        <FormattedMessage
+          id="app.tutorial.slide.withers"
+          defaultMessage=" kg and his withers height was 3,96 meters. That made him a meter higher than an average african elephant!"
+          description="Tut4 withers"
+        />
       </Text>
       <Flex mt="auto" justifyContent="space-between" width="100%">
-        <Button
-          onClick={handleClickHome}
-          backgroundColor={'#414141'}
-          sx={{ flex: 1 }}
-          mr="1"
-        >
-          Domů
-        </Button>
-        <Button onClick={handleClickNext} sx={{ flex: 5 }}>
-          Další
-        </Button>
+        <HomeButton />
+        <NextButton handleNextStep={handleNextStep} />
       </Flex>
     </Container>
   );
