@@ -1,33 +1,36 @@
 import React from 'react';
-import { Button, Box, Text, Flex } from 'rebass';
+import { Box, Text, Flex } from 'rebass';
 import NumberFormat from 'react-number-format';
 import { Container } from '../Container';
 import { TutorialHeader } from '../TutorialHeader';
 import { SlideProps } from './types';
 import { ScoreChart } from '../ScoreChart';
-import { useHistory } from 'react-router';
+import { HomeButton } from './HomeButton';
+import { NextButton } from './NextButton';
+import { FormattedMessage } from 'react-intl';
 
 const Slide5 = ({ handleNextStep, currentTip }: SlideProps) => {
   const correctAnswer = 11000;
   const previousTips = [8300, 9500, 15000];
-  const history = useHistory();
-  const handleClickHome = () => {
-    history.push('/');
-  };
-  const handleClickNext = () => {
-    handleNextStep();
-  };
   return (
     <Container>
       <TutorialHeader>
         <Text fontSize={[3, 4, 5]} color="secondary" textAlign="center" p={3}>
-          Tvůj tip:{' '}
+          <FormattedMessage
+            id="app.tipeval"
+            defaultMessage="Your tip: "
+            description="Tipeval"
+          />
           <NumberFormat
             value={currentTip}
             displayType={'text'}
             thousandSeparator={' '}
           />{' '}
-          kg
+          <FormattedMessage
+            id="app.kgunit"
+            defaultMessage="kg"
+            description="kgunit"
+          />
         </Text>
       </TutorialHeader>
       <Box width="100%" height="200px">
@@ -46,31 +49,42 @@ const Slide5 = ({ handleNextStep, currentTip }: SlideProps) => {
         py={4}
         px={3}
       >
-        Na tomto grafu můžeš vidět{' '}
+        <FormattedMessage
+          id="app.tutorial.slide.chart"
+          defaultMessage="In this chart you can see "
+          description="Tut5 chart"
+        />
         <Text color="accent" as="span">
-          svou odpověď
+          <FormattedMessage
+            id="app.tutorial.slide.answer"
+            defaultMessage="your answer"
+            description="Tut5 answer"
+          />
         </Text>
         ,{' '}
         <Text color="primary" as="span">
-          uvedené nápovědy
+          <FormattedMessage
+            id="app.tutorial.slide.prevtips"
+            defaultMessage="previous tips"
+            description="Tut5 prevtips"
+          />
         </Text>{' '}
-        a 
+        <FormattedMessage
+          id="app.tutorial.slide.and"
+          defaultMessage="and "
+          description="Tut5 and"
+        />
         <Text color="white" as="span">
-          správnou odpověď.
+          <FormattedMessage
+            id="app.tutorial.slide.correct"
+            defaultMessage="the correct answer."
+            description="Tut5 correct answer"
+          />
         </Text>
       </Text>
       <Flex mt="auto" justifyContent="space-between" width="100%">
-        <Button
-          onClick={handleClickHome}
-          backgroundColor={'#414141'}
-          sx={{ flex: 1 }}
-          mr="1"
-        >
-          Domů
-        </Button>
-        <Button onClick={handleClickNext} sx={{ flex: 5 }}>
-          Další
-        </Button>
+        <HomeButton />
+        <NextButton handleNextStep={handleNextStep} />
       </Flex>
     </Container>
   );
