@@ -1,33 +1,36 @@
 import React from 'react';
-import { Button, Text, Box, Flex } from 'rebass';
+import { Text, Box, Flex } from 'rebass';
 import { getScore } from '../../helpers';
 import { Container } from '../Container';
 import { TutorialHeader } from '../TutorialHeader';
 import { SlideProps } from './types';
 import { ScoreChart } from '../ScoreChart';
-import { useHistory } from 'react-router';
+import { HomeButton } from './HomeButton';
+import { NextButton } from './NextButton';
+import { FormattedMessage } from 'react-intl';
 
 const Slide6 = ({ handleNextStep, currentTip }: SlideProps) => {
   const correctAnswer = 11000;
-  const history = useHistory();
-  const handleClickHome = () => {
-    history.push('/');
-  };
-  const handleClickNext = () => {
-    handleNextStep();
-  };
   return (
     <Container>
       <TutorialHeader>
         <Text fontSize={[3, 4, 5]} color="secondary" textAlign="center" p={3}>
-          Za svou odpověď bys dostal*a{' '}
+          <FormattedMessage
+            id="app.tutorial.slide.youdget"
+            defaultMessage="For your answer you would get "
+            description="Tut6 you'd get"
+          />
           <Text color="accent" as="span">
             {' '}
             {typeof currentTip !== 'undefined'
               ? getScore(currentTip, 11000).toFixed(2)
               : 0}{' '}
           </Text>
-          bodu.
+          <FormattedMessage
+            id="app.tutorial.slide.points"
+            defaultMessage="points."
+            description="Tut6 points"
+          />
         </Text>
       </TutorialHeader>
       <Box width="100%" height="200px">
@@ -42,20 +45,15 @@ const Slide6 = ({ handleNextStep, currentTip }: SlideProps) => {
         py={4}
         px={3}
       >
-        Body získáš, pokud se trefíš mezi 50 - 200 % hodnoty správné odpovědi.
+        <FormattedMessage
+          id="app.tutorial.slide.whenpoints"
+          defaultMessage="You get points if you hit between 50 - 200% of the value of the correct answer."
+          description="Tut6 whenpoints"
+        />
       </Text>
       <Flex mt="auto" justifyContent="space-between" width="100%">
-        <Button
-          onClick={handleClickHome}
-          backgroundColor={'#414141'}
-          sx={{ flex: 1 }}
-          mr="1"
-        >
-          Domů
-        </Button>
-        <Button onClick={handleClickNext} sx={{ flex: 5 }}>
-          Další
-        </Button>
+        <HomeButton />
+        <NextButton handleNextStep={handleNextStep} />
       </Flex>
     </Container>
   );
