@@ -2,41 +2,41 @@ import React from 'react';
 import Helmet from 'react-helmet';
 import { Text } from 'rebass';
 import { Container } from './../../components/Container';
+import { FormattedMessage, useIntl } from 'react-intl';
 
-const Consent = () => (
-  <Container>
-    <Helmet title="Souhlas s podmínkami"></Helmet>
-    <Text color="secondary" my="6">
-      Tímto uděluji souhlas Bc. Pavlíně Hillerové (hillerovap@natur.cuni.cz),
-      která je správcem osobních údajů, se zpracováním svých osobních údajů
-      v rozsahu
-      <br />
-      -  emailová adresa
-      <br />
-      -  osobních údajů, které vyplynou z odpovědí na v dotaznících uvedené
-      otázky
-      <br />
-      za účelem studie kulturního přenosu pro diplomovou práci.
-      <br />
-      <br />
-      Beru na vědomí, že
-      <br />
-      -  mi budou zasílány informace na uvedenou e-mailovou adresu,
-      <br />
-      -  mám práva: o požádat o informaci, jaké osobní údaje jsou o mé osobě
-      zpracovávány,
-      <br />
-      -  požadovat opravu osobních údajů, pokud jsou neplatné nebo zastaralé,
-      <br />
-      -  požadovat, aby nebyly moje osobní údaje zpracovávány do doby, než bude
-      vyřešena oprávněnost výše uvedených požadavků,
-      <br />
-      -  podat stížnost u dozorového úřadu,
-      <br />
-      -  souhlas kdykoli odvolat zrušením svého profilu zasláním e-mailu na
-      adresu hillerovap@natur.cuni.cz.
-    </Text>
-  </Container>
-);
+const Consent = () => {
+  const intl = useIntl();
+  const helmet = intl.formatMessage({
+    id: 'app.consenthelmet',
+    defaultMessage: `Consent to Terms`,
+    description: 'Consent helmet',
+  });
+  return (
+    <Container>
+      <Helmet title={helmet}></Helmet>
+      <Text color="secondary" my="6">
+        <FormattedMessage
+          id="app.consent"
+          defaultMessage="I hereby consent to Mgr. Pavlína Hillerová (hillerovap@natur.cuni.cz) processing my personal data, specifically: {lineBreak}
+          - my email address {lineBreak}
+          - personal information derived from my responses to questions in the questionnaires, {lineBreak}
+        for the purposes of a study on cultural transmission. {lineBreak}
+        {lineBreak}
+        I understand that: {lineBreak}
+        - I will receive information at the email address provided, {lineBreak}
+        - I have the right to request information on what personal data is being processed about me, {lineBreak}
+        - I can request corrections to my personal data if it is inaccurate or outdated, {lineBreak}
+        - I can request that my personal data not be processed until the validity of the above requests is resolved, {lineBreak}
+        - I have the right to lodge a complaint with the supervisory authority, {lineBreak}
+        - I may withdraw my consent at any time by requesting deletion of my profile via email to hillerovap@natur.cuni.cz."
+          description="Consent"
+          values={{
+            lineBreak: <br />,
+          }}
+        />
+      </Text>
+    </Container>
+  );
+};
 
 export { Consent };
