@@ -1,8 +1,10 @@
 import { useMutation, useQuery } from '@apollo/client';
 import React, { useState } from 'react';
 import Helmet from 'react-helmet';
-import { Redirect } from 'react-router-dom';
-import { Box, Button, Flex, Heading, Text } from 'rebass';
+import { FormattedMessage } from 'react-intl';
+import { redirect } from 'react-router-dom';
+import { Box, Flex, Heading, Text } from 'rebass';
+import Button from '@mui/material/Button';
 import { BackButton, Container, Spinner } from '../../components';
 import {
   EXPORT_MUTATION,
@@ -10,6 +12,8 @@ import {
   ONLINE_STATS_QUERY,
 } from '../../gql';
 import { User, UserRole } from '../../types';
+
+const buttonStyles = { width: '50%', my: 2 };
 
 const consoleStyle = {
   px: 2,
@@ -71,7 +75,7 @@ const Dashboard = ({ user }: DashboardProps) => {
   };
 
   if (!user || user.role !== UserRole.admin) {
-    return <Redirect to="/" />;
+    redirect('/');
   }
 
   if (loading) {
@@ -93,29 +97,49 @@ const Dashboard = ({ user }: DashboardProps) => {
     <Container>
       <Helmet title="Admin dashboard"></Helmet>
       <Heading color="primary" my={4}>
-        Dashboard
+        <FormattedMessage
+          id="app.dashboard.menu.dasboard"
+          defaultMessage="Dashboard"
+          description="Dashboard"
+        />
       </Heading>
       <Flex flexDirection="column">
         <Heading fontSize={3} color="white" mb={3}>
-          Actions
+          <FormattedMessage
+            id="app.dashboard.menu.actions"
+            defaultMessage="Actions"
+            description="Actions button"
+          />
         </Heading>
         <Flex justifyContent="space-between" alignItems="center">
           <Flex flex={1} flexDirection="column" mr={1}>
             <Text color="white" fontFamily="Tahoma">
-              Export .csv to S3
+              <FormattedMessage
+                id="app.dashboard.menu.export"
+                defaultMessage="Export .csv to S3"
+                description="Export button"
+              />
             </Text>
-            <Button width="50%" my={2} onClick={handleClickExport}>
+            <Button
+              variant="contained"
+              sx={buttonStyles}
+              onClick={handleClickExport}
+            >
               {exportLoading ? '... generating' : 'Download'}
             </Button>
           </Flex>
           <Flex flex={1} flexDirection="column">
             <Text color="white" fontFamily="Tahoma">
-              Import from Spreadsheet
+              <FormattedMessage
+                id="app.dashboard.menu.import"
+                defaultMessage="Import from Spreadsheet"
+                description="Import button"
+              />
             </Text>
             <Button
+              variant="contained"
               disabled={importLoading}
-              width="50%"
-              my={2}
+              sx={buttonStyles}
               onClick={handleClickImport}
             >
               {importLoading ? '... importing' : 'Import'}
@@ -130,16 +154,30 @@ const Dashboard = ({ user }: DashboardProps) => {
           ))}
         </Box>
         <Heading fontSize={3} color="white" mb={3}>
-          Stats
+          <FormattedMessage
+            id="app.dashboard.menu.stats"
+            defaultMessage="Stats"
+            description="Stats button"
+          />
         </Heading>
         <Flex flexDirection="column">
-          <Text color="white" fontFamily="Tahoma" mb={3}>
-            <b>Online users playing: </b>
+          <Text color="white" fontWeight="bold" fontFamily="Tahoma" mb={3}>
+            <FormattedMessage
+              id="app.dashboard.menu.online"
+              defaultMessage="Online users playing: "
+              description="Online button"
+            />
             {data.getOnlineStats.online}
           </Text>
         </Flex>
       </Flex>
-      <BackButton>domů</BackButton>
+      <BackButton>
+        <FormattedMessage
+          id="app.dashboard.footer.home"
+          defaultMessage="Home"
+          description="Home button"
+        />
+      </BackButton>
     </Container>
   );
 };
