@@ -1,7 +1,8 @@
 import React, { useRef, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { Input, Label, Radio } from '@rebass/forms';
-import { Flex, Button, Text, Box } from 'rebass';
+import { Flex, Text, Box } from 'rebass';
+import { FormattedMessage, useIntl } from 'react-intl';
 import * as Yup from 'yup';
 import {
   inputStyles,
@@ -15,6 +16,8 @@ import {
 import { useMutation } from '@apollo/client';
 import { AuthQueryName, UPDATE_USER_MUTATION } from '../../../gql';
 import { User } from '../../../types';
+import Button from '@mui/material/Button';
+import { validationMessages } from './messages';
 
 enum Genders {
   man = 'man',
@@ -29,51 +32,63 @@ const getOtherGender = (gender: string) =>
 const otherGenderInputStyles = { ...inputStyles, mt: 2 };
 const radioLabelStyles = { ...labelStyles, mb: 2 };
 
-const validationSchema = Yup.object().shape({
-  email: Yup.string()
-    .notRequired()
-    .matches(emailRegex, 'Prosím, zadej platný e-mail')
-    .nullable()
-    .transform((value) => (!!value ? value : undefined)),
-  oldPassword: Yup.mixed()
-    .when('newPassword', {
-      is: (val) => val !== '' && val != null,
-      then: Yup.string().required(
-        'Pro změnu hesla musíš vyplnit aktuální heslo'
-      ),
-      otherwise: Yup.mixed().notRequired(),
-    })
-    .strict(),
-  confirmNewPassword: Yup.mixed()
-    .when('newPassword', {
-      is: (val) => val !== '' && val != null,
-      then: Yup.string()
-        .oneOf([Yup.ref('newPassword')], 'Hesla se neshodují')
-        .required(),
-      otherwise: Yup.mixed().notRequired(),
-    })
-    .strict(),
-  newPassword: Yup.string()
-    .notRequired()
-    .min(
-      MIN_PASSWORD_LENGTH,
-      `Heslo musí mít alespoň ${MIN_PASSWORD_LENGTH} znaků`
-    )
-    .nullable()
-    .transform((value) => (!!value ? value : undefined)),
-  age: Yup.number()
-    .notRequired()
-    .min(18)
-    .max(120)
-    .nullable()
-    .transform((value) => (!!value ? value : undefined)),
-  gender: Yup.string()
-    .notRequired()
-    .nullable()
-    .transform((value) => (!!value ? value : undefined)),
-});
-
 const Settings = ({ user }: { user: User | null }) => {
+  const intl = useIntl();
+  const placeholder = intl.formatMessage({
+    id: 'app.settings.menu.specifygender',
+    defaultMessage: 'specify (if you like)',
+    description: 'specify gender',
+  });
+  const validationSchema = Yup.object().shape({
+    email: Yup.string()
+      .notRequired()
+      .matches(emailRegex, intl.formatMessage(validationMessages.emailInvalid))
+      .nullable()
+      .transform((value) => (!!value ? value : undefined)),
+    oldPassword: Yup.mixed()
+      .when('newPassword', ([newPassword, schema]) => {
+        return newPassword !== '' && newPassword != null
+          ? schema
+              .string()
+              .required(intl.formatMessage(validationMessages.passwordCurrent))
+          : schema.mixed().notRequired();
+      })
+      .strict(),
+    confirmNewPassword: Yup.mixed()
+      .when('newPassword', ([confirmNewPassword, schema]) => {
+        return confirmNewPassword !== '' && confirmNewPassword != null
+          ? schema
+              .string()
+              .oneOf(
+                [Yup.ref('newPassword')],
+                intl.formatMessage(validationMessages.passwordNotMatch)
+              )
+              .required()
+          : schema.mixed().notRequired();
+      })
+      .strict(),
+    newPassword: Yup.string()
+      .notRequired()
+      .min(
+        MIN_PASSWORD_LENGTH,
+        intl.formatMessage(validationMessages.passwordMin, {
+          min: MIN_PASSWORD_LENGTH,
+        })
+      )
+      .nullable()
+      .transform((value) => (!!value ? value : undefined)),
+    age: Yup.number()
+      .notRequired()
+      .min(18)
+      .max(120)
+      .nullable()
+      .transform((value) => (!!value ? value : undefined)),
+    gender: Yup.string()
+      .notRequired()
+      .nullable()
+      .transform((value) => (!!value ? value : undefined)),
+  });
+
   const otherGenderInputRef = useRef();
   const [log, setLog] = useState<boolean | undefined>();
   const [errors, setErrors] = useState<Array<{ error: string }>>([]);
@@ -152,62 +167,27 @@ const Settings = ({ user }: { user: User | null }) => {
       flexDirection="column"
     >
       <Text color="secondary" my={3}>
-        Všechny údaje budou před vyhodnocením anonymizovány.
+        <FormattedMessage
+          id="app.settings.menu.anon"
+          defaultMessage="All data will be anonymized before evaluation."
+          description="Anon text"
+        />
       </Text>
-
-      <Text color="secondary" fontSize={3} mt={2} mb={3} px={3}>
-        Nastavení účtu
-      </Text>
-      <Label htmlFor="email" sx={labelStyles}>
-        E-mail
-      </Label>
-      <Input
-        id="email"
-        type="text"
-        placeholder="e-mail"
-        mb={2}
-        sx={inputStyles}
-        {...register('email')}
-      />
-      <Label htmlFor="oldPassword" sx={labelStyles}>
-        Aktuální heslo
-      </Label>
-      <Input
-        autoComplete="new-password"
-        id="oldPassword"
-        type="password"
-        mb={2}
-        sx={inputStyles}
-        {...register('oldPassword')}
-      />
-      <Label htmlFor="newPassword" sx={labelStyles}>
-        Nové heslo
-      </Label>
-      <Input
-        id="newPassword"
-        type="password"
-        mb={2}
-        sx={inputStyles}
-        {...register('newPassword')}
-      />
-      <Label htmlFor="confirmNewPassword" sx={labelStyles}>
-        Nové heslo znovu
-      </Label>
-      <Input
-        id="confirmNewPassword"
-        type="password"
-        mb={2}
-        sx={inputStyles}
-        {...register('confirmNewPassword')}
-      />
-
-      <Text color="secondary" fontSize={3} mt={4} px={3}>
-        Nepovinné údaje
+      <Text color="secondary" fontSize={3} mt={1} px={3}>
+        <FormattedMessage
+          id="app.settings.menu.otherinfo"
+          defaultMessage="Additional (voluntary) info"
+          description="Otherinfo text"
+        />
       </Text>
 
       <Box my={3}>
         <Label htmlFor="gender" sx={radioLabelStyles}>
-          Jsem
+          <FormattedMessage
+            id="app.settings.menu.gender"
+            defaultMessage="I am..."
+            description="Gender label"
+          />
         </Label>
         <Label color="white">
           <Radio
@@ -216,15 +196,27 @@ const Settings = ({ user }: { user: User | null }) => {
             id="woman"
             value="woman"
           />
-          Žena
+          <FormattedMessage
+            id="app.settings.menu.woman"
+            defaultMessage="Woman"
+            description="Woman label"
+          />
         </Label>
         <Label color="white">
           <Radio {...register('gender')} name="gender" id="man" value="man" />
-          Muž
+          <FormattedMessage
+            id="app.settings.menu.man"
+            defaultMessage="Man"
+            description="Man label"
+          />
         </Label>
         <Label color="white">
           <Radio {...register('gender')} name="gender" id="enby" value="enby" />
-          Nebinární
+          <FormattedMessage
+            id="app.settings.menu.nonbinary"
+            defaultMessage="Nonbinary"
+            description="Nonbinary label"
+          />
         </Label>
         <Label color="white">
           <Radio
@@ -233,22 +225,31 @@ const Settings = ({ user }: { user: User | null }) => {
             id="other"
             value="other"
           />
-          Jiné (můžeš uvést):
+          <FormattedMessage
+            id="app.settings.menu.othergender"
+            defaultMessage="Other (feel free to specify):"
+            description="Othergender label"
+          />
         </Label>
         {watchGender === 'other' && (
           <Input
             ref={otherGenderInputRef}
-            placeholder="uveď, prosím"
+            placeholder={placeholder}
             mb="2"
             defaultValue={user.gender}
             sx={otherGenderInputStyles}
           />
         )}
       </Box>
-      <Label htmlFor="email" sx={labelStyles}>
-        Věk
+      <Label htmlFor="age" sx={labelStyles}>
+        <FormattedMessage
+          id="app.settings.menu.age"
+          defaultMessage="Age:"
+          description="Age label"
+        />
       </Label>
       <Input
+        id="age"
         type="number"
         name="age"
         min={18}
@@ -266,6 +267,71 @@ const Settings = ({ user }: { user: User | null }) => {
           ))}
         </Box>
       )}
+      <Text color="secondary" fontSize={3} mt={4} mb={3} px={3}>
+        <FormattedMessage
+          id="app.settings.menu.settings"
+          defaultMessage="Account settings"
+          description="Settings header"
+        />
+      </Text>
+      <Label htmlFor="email" sx={labelStyles}>
+        <FormattedMessage
+          id="app.settings.menu.email"
+          defaultMessage="E-mail"
+          description="Email label"
+        />
+      </Label>
+      <Input
+        id="email"
+        type="text"
+        placeholder="e-mail"
+        mb={2}
+        sx={inputStyles}
+        {...register('email')}
+      />
+      <Label htmlFor="oldPassword" sx={labelStyles}>
+        <FormattedMessage
+          id="app.settings.menu.passwordcur"
+          defaultMessage="Current password"
+          description="PasswordCur label"
+        />
+      </Label>
+      <Input
+        autoComplete="new-password"
+        id="oldPassword"
+        type="password"
+        mb={2}
+        sx={inputStyles}
+        {...register('oldPassword')}
+      />
+      <Label htmlFor="newPassword" sx={labelStyles}>
+        <FormattedMessage
+          id="app.settings.menu.passwordnew"
+          defaultMessage="New password"
+          description="PasswordNew label"
+        />
+      </Label>
+      <Input
+        id="newPassword"
+        type="password"
+        mb={2}
+        sx={inputStyles}
+        {...register('newPassword')}
+      />
+      <Label htmlFor="confirmNewPassword" sx={labelStyles}>
+        <FormattedMessage
+          id="app.settings.menu.passwordnew2"
+          defaultMessage="Reenter new password"
+          description="PasswordNew2 label"
+        />
+      </Label>
+      <Input
+        id="confirmNewPassword"
+        type="password"
+        mb={2}
+        sx={inputStyles}
+        {...register('confirmNewPassword')}
+      />
 
       <Flex mt="auto" flexDirection="column" width="100%" mb="2">
         {log === true && (
@@ -276,10 +342,22 @@ const Settings = ({ user }: { user: User | null }) => {
             backgroundColor="#15de46"
             alignItems="center"
           >
-            <Text>✓ Nastavení uloženo</Text>
+            <Text>
+              <FormattedMessage
+                id="app.settings.menu.saved"
+                defaultMessage="✓ Settings updated"
+                description="Settingsupdated label"
+              />
+            </Text>
           </Flex>
         )}
-        <Button sx={{ flex: 1, color: 'white' }}>Uložit</Button>
+        <Button variant="contained" sx={{ flex: 1, color: 'white' }}>
+          <FormattedMessage
+            id="app.settings.menu.save"
+            defaultMessage="Save"
+            description="Save label"
+          />
+        </Button>
       </Flex>
     </Flex>
   );
