@@ -4,14 +4,26 @@ import { HelmetData } from 'react-helmet';
 interface DocumentProps {
   content: string;
   helmet: HelmetData;
-  js: string;
   state: any;
   css: React.ReactNode;
+  linkTags: React.ReactNode;
+  styleTags: React.ReactNode;
+  scriptTags: React.ReactNode;
+  initialLanguage: string;
 }
 
-export function Document({ content, css, helmet, js, state }: DocumentProps) {
+export function Document({
+  content,
+  css,
+  helmet,
+  state,
+  linkTags,
+  styleTags,
+  scriptTags,
+  initialLanguage,
+}: DocumentProps) {
   return (
-    <html lang="sk">
+    <html lang={initialLanguage}>
       <head>
         <meta charSet="UTF-8" />
         <meta name="google" content="notranslate" />
@@ -22,16 +34,14 @@ export function Document({ content, css, helmet, js, state }: DocumentProps) {
         {helmet.title.toComponent()}
         {helmet.meta.toComponent()}
         {helmet.link.toComponent()}
-        <script
-          src={js}
-          defer
-          crossOrigin={(process.env.NODE_ENV !== 'production').toString()}
-        />
+        {css}
+        {linkTags}
+        {styleTags}
         <link rel="preconnect" href="https://fonts.googleapis.com" />
         <link
           rel="preconnect"
           href="https://fonts.gstatic.com"
-          crossOrigin="true"
+          crossOrigin="anonymous"
         />
         <link
           href="https://fonts.googleapis.com/css2?family=Raleway:wght@900&display=swap"
@@ -41,7 +51,7 @@ export function Document({ content, css, helmet, js, state }: DocumentProps) {
         <link
           rel="preconnect"
           href="https://fonts.gstatic.com"
-          crossOrigin="true"
+          crossOrigin="anonymous"
         />
         <link
           href="https://fonts.googleapis.com/css2?family=Jost:wght@300;400;700&display=swap"
@@ -51,6 +61,12 @@ export function Document({ content, css, helmet, js, state }: DocumentProps) {
       </head>
       <body>
         <div id="root" dangerouslySetInnerHTML={{ __html: content }} />
+        {scriptTags}
+        <script
+          dangerouslySetInnerHTML={{
+            __html: `window.__INITIAL_LANGUAGE__="${initialLanguage}";`,
+          }}
+        />
         <script
           dangerouslySetInnerHTML={{
             __html: `window.__D=${JSON.stringify(state).replace(
